firmware: Remove SNP and associate types
