+++ src/firmware/types.rs
@@ -5,13 +5,52 @@
 #[cfg(target_os = "linux")]
 use super::*;
 
-pub use crate::firmware::linux::types::{PlatformStatusFlags, SnpConfig, TcbVersion};
-
-use crate::firmware::linux::_4K_PAGE;
-
 use std::{error, fmt::Debug, io};
 
-use serde::{Deserialize, Serialize};
+bitflags::bitflags! {
+    /// The platform's status flags.
+    #[derive(Default)]
+    pub struct PlatformStatusFlags: u32 {
+        /// If set, this platform is owned. Otherwise, it is self-owned.
+        const OWNED           = 1 << 0;
+
+        /// If set, encrypted state functionality is present.
+        const ENCRYPTED_STATE = 1 << 8;
+    }
+}
+
+/// TcbVersion represents the version of the firmware.
+///
+/// (Chapter 2.2; Table 3)
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
+#[repr(C)]
+pub struct TcbVersion {
+    /// Current bootloader version.
+    /// SVN of PSP bootloader.
+    pub bootloader: u8,
+    /// Current PSP OS version.
+    /// SVN of PSP operating system.
+    pub tee: u8,
+    _reserved: [u8; 4],
+    /// Version of the SNP firmware.
+    /// Security Version Number (SVN) of SNP firmware.
+    pub snp: u8,
+    /// Lowest current patch level of all the cores.
+    pub microcode: u8,
+}
+
+impl TcbVersion {
+    /// Creates a new isntance of a TcbVersion
+    pub fn new(bootloader: u8, tee: u8, snp: u8, microcode: u8) -> Self {
+        Self {
+            bootloader,
+            tee,
+            snp,
+            microcode,
+            _reserved: Default::default(),
+        }
+    }
+}
 
 /// There are a number of error conditions that can occur between this
 /// layer all the way down to the SEV platform. Most of these cases have
@@ -32,9 +71,6 @@ pub enum UserApiError {
     /// Firmware related errors.
     FirmwareError(Error),
 
-    /// User API related errors.
-    ApiError(SnpCertError),
-
     /// Uuid parsing errors.
     UuidError(uuid::Error),
 }
@@ -42,7 +78,6 @@ pub enum UserApiError {
 impl error::Error for UserApiError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
-            UserApiError::ApiError(uapi_error) => Some(uapi_error),
             UserApiError::FirmwareError(firmware_error) => Some(firmware_error),
             UserApiError::UuidError(uuid_error) => Some(uuid_error),
         }
@@ -53,7 +88,6 @@ impl std::fmt::Display for UserApiError {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         let err_msg: String = match self {
             UserApiError::FirmwareError(error) => format!("Firmware Error Encountered: {error}"),
-            UserApiError::ApiError(error) => format!("Certificate Error Encountered: {error}"),
             UserApiError::UuidError(error) => format!("UUID Error Encountered: {error}"),
         };
         write!(f, "{err_msg}")
@@ -77,47 +111,6 @@ impl std::convert::From<std::io::Error> for UserApiError {
         UserApiError::FirmwareError(Error::IoError(io_error))
     }
 }
-
-impl std::convert::From<SnpCertError> for UserApiError {
-    fn from(cert_error: SnpCertError) -> Self {
-        UserApiError::ApiError(cert_error)
-    }
-}
-
-#[derive(Debug)]
-/// Errors which may be encountered through misuse of the User API.
-pub enum SnpCertError {
-    /// Malformed GUID.
-    InvalidGUID,
-
-    /// Malformed Page Alignment
-    PageMisalignment,
-
-    /// Invalid Buffer Size
-    BufferOverflow,
-
-    /// Unknown Error.
-    UnknownError,
-}
-
-impl std::fmt::Display for SnpCertError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            SnpCertError::InvalidGUID => write!(f, "Invalid GUID provided in certificate chain."),
-            SnpCertError::PageMisalignment => {
-                write!(f, "Certificate Buffer not aligned with 4K Pages.")
-            }
-            SnpCertError::BufferOverflow => {
-                write!(f, "Buffer overflow prevented: Bytes provided exceed space allocated for the buffer provided.")
-            }
-            SnpCertError::UnknownError => {
-                write!(f, "Unknown Error encountered within the certificate chain.")
-            }
-        }
-    }
-}
-
-impl error::Error for SnpCertError {}
 
 /// Error conditions returned by the SEV platform or by layers above it
 /// (i.e., the Linux kernel).
@@ -391,207 +384,5 @@ impl std::fmt::Display for Identifier {
         }
 
         Ok(())
-    }
-}
-
-/// Information regarding the SEV-SNP platform's TCB version.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SnpTcbStatus {
-    /// Installed TCB version.
-    pub platform_version: TcbVersion,
-
-    /// Reported TCB version.
-    pub reported_version: TcbVersion,
-}
-
-/// Information regarding the SEV-SNP platform's current status.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SnpStatus {
-    /// The build number.
-    pub build: SnpBuild,
-
-    /// The platform's current state.
-    pub state: State,
-
-    /// IsRmpInitiailzied
-    pub is_rmp_init: bool,
-
-    /// MaskChipId
-    pub mask_chip_id: bool,
-
-    /// The number of valid guests supervised by this platform.
-    pub guests: u32,
-
-    /// TCB status.
-    pub tcb: SnpTcbStatus,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-#[repr(C)]
-/// Certificates which are accepted for [`CertTableEntry`]
-pub enum SnpCertType {
-    /// AMD Root Signing Key (ARK) certificate
-    ARK,
-
-    /// AMD SEV Signing Key (ASK) certificate
-    ASK,
-
-    /// Versioned Chip Endorsement Key (VCEK) certificate
-    VCEK,
-
-    /// Other (Specify GUID)
-    OTHER(String),
-
-    /// Empty or closing entry for the CertTable
-    Empty,
-}
-
-impl SnpCertType {
-    /// Create a certificate from the specified GUID. Any unexpected matches
-    /// produce an [`SnpCertType::OTHER`] type from the guid provided.
-    fn from_guid(guid: &str) -> Self {
-        match guid {
-            "c0b406a4-a803-4952-9743-3fb6014cd0ae" => SnpCertType::ARK,
-            "4ab7b379-bbac-4fe4-a02f-05aef327c782" => SnpCertType::ASK,
-            "63da758d-e664-4564-adc5-f4b93be8accd" => SnpCertType::VCEK,
-            "00000000-0000-0000-0000-000000000000" => SnpCertType::Empty,
-            guid => SnpCertType::OTHER(guid.to_string()),
-        }
-    }
-
-    /// Retreive the GUID from the [`SnpCertType`].
-    fn guid(&self) -> String {
-        match self {
-            SnpCertType::ARK => "c0b406a4-a803-4952-9743-3fb6014cd0ae",
-            SnpCertType::ASK => "4ab7b379-bbac-4fe4-a02f-05aef327c782",
-            SnpCertType::VCEK => "63da758d-e664-4564-adc5-f4b93be8accd",
-            SnpCertType::Empty => "00000000-0000-0000-0000-000000000000",
-            SnpCertType::OTHER(guid) => guid,
-        }
-        .to_string()
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-#[repr(C)]
-/// An entry with information regarding a specific certificate.
-pub struct CertTableEntry {
-    /// A Specificy certificate type.
-    pub cert_type: SnpCertType,
-
-    /// The raw data of the certificate.
-    pub data: Vec<u8>,
-}
-
-impl CertTableEntry {
-    /// Façade for retreiving the GUID for the Entry.
-    pub fn guid(&self) -> String {
-        self.cert_type.guid()
-    }
-
-    /// Get an immutable reference to the data stored in the entry.
-    pub fn data(&self) -> &[u8] {
-        self.data.as_slice()
-    }
-
-    /// Generates a certificate from the str GUID and data provided.
-    pub fn from_guid(guid: &str, data: Vec<u8>) -> Self {
-        Self {
-            cert_type: SnpCertType::from_guid(guid),
-            data,
-        }
-    }
-
-    /// Generates a certificate from the SnpCertType and data provided.
-    pub fn new(cert_type: SnpCertType, data: Vec<u8>) -> Self {
-        Self { cert_type, data }
-    }
-}
-
-/// Rust-friendly instance of the SNP Extended Configuration.
-/// It may be used either to fetch or set the configuration.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct SnpExtConfig {
-    /// SET:
-    ///     Address of the SnpConfig or 0 when reported_tcb does not need
-    ///     to be updated.
-    ///
-    /// GET:
-    ///     Address of the SnpConfig or 0 when reported_tcb should not be
-    ///     fetched.
-    pub config: Option<SnpConfig>,
-
-    /// SET:
-    ///     Address of extended guest request certificate chain or None when
-    ///     previous certificate should be removed on SNP_SET_EXT_CONFIG.
-    ///
-    /// GET:
-    ///     Address of extended guest request certificate chain or None when
-    ///     certificate should not be fetched.
-    pub certs: Option<Vec<CertTableEntry>>,
-
-    /// SET:
-    ///     Length of the certificates.
-    ///
-    /// GET:
-    ///     Length of the buffer which will hold the fetched certificates.
-    pub certs_len: u32,
-}
-
-/// Used to round certificate buffers to 4K page alignment.
-fn round_to_whole_pages(size: usize) -> usize {
-    match size % _4K_PAGE {
-        0 => size,
-        rem => size + (_4K_PAGE - rem),
-    }
-}
-
-impl SnpExtConfig {
-    /// Used to update the PSP with the cerificates provided.
-    pub fn update_certs_only(certificates: Vec<CertTableEntry>) -> Result<Self, SnpCertError> {
-        let certs_length: usize = certificates.iter().map(|entry| entry.data().len()).sum();
-        let certs_len: u32 = round_to_whole_pages(certs_length) as u32;
-
-        Ok(Self {
-            config: None,
-            certs: Some(certificates),
-            certs_len,
-        })
-    }
-}
-
-#[cfg(test)]
-mod test {
-
-    use super::{CertTableEntry, SnpCertType, SnpConfig, SnpExtConfig};
-    use crate::firmware::linux::types::TcbVersion;
-
-    fn build_ext_config() -> SnpExtConfig {
-        let test_cfg: SnpConfig = SnpConfig::new(TcbVersion::new(2, 0, 6, 39), 31);
-
-        let cert_table: Vec<CertTableEntry> =
-            vec![CertTableEntry::new(SnpCertType::ARK, vec![1; 28])];
-
-        SnpExtConfig {
-            config: Some(test_cfg),
-            certs: Some(cert_table),
-            certs_len: 4096,
-        }
-    }
-
-    #[test]
-    fn snp_ext_config_get_config() {
-        let expected_data: SnpConfig = SnpConfig::new(TcbVersion::new(2, 0, 6, 39), 31);
-        let cfg: SnpExtConfig = build_ext_config();
-        assert_eq!(cfg.config, Some(expected_data));
-    }
-
-    #[test]
-    fn snp_ext_config_get_certs() {
-        let cert_table: Vec<CertTableEntry> =
-            vec![CertTableEntry::new(SnpCertType::ARK, vec![1; 28])];
-
-        let cfg: SnpExtConfig = build_ext_config();
-        assert_eq!(cfg.certs, Some(cert_table));
     }
 }
